requirements.txt
autopep8==1.5.7
click==8.0.1
et-xmlfile==1.1.0
Faker==8.12.1
Flask==2.0.1
Flask-MQTT==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
openpyxl==3.0.9
paho-mqtt==1.5.1
pycodestyle==2.7.0
python-dateutil==2.8.2
PyYAML==5.4.1
six==1.16.0
text-unidecode==1.3
toml==0.10.2
Werkzeug==2.0.1
